Use all item names in datalist for autocompletion
Before I was filtering the list of items down to just those that
matched the current search string. However, the matching is
automatically handled by the browser. Now we just populate the
datalist with the names of all the items in the list and let the
browser do all the clever stuff.