Add a submit form button
Added this button to make the form more accessible and to make it
functional on mobile. Swapped row -> form-row to get the spacing a
little tighter and reduced the number of columns for the name input
to make space for the new button. I don't love the spacing of the form
in its current style, but I think the added functionality/accessibilty
is worth while.